Adding more routers.

+++ src/controllers/api/categories.ts
@@ -3,24 +3,28 @@ const NOT_FOUND = 'Category not found.';
 const ALL_REQUIRED = 'Title and Order fields are required.';
 const SOME_REQUIRED = 'Title or Order field is required.';
 const INVALID = 'One or more fields are invalid.';
+const SERVER_ERROR = 'Server error.'
 
 import {Router, json} from 'express';
 import apiKey from '../../middleware/api-key';
 import isAdmin from '../../middleware/is-admin';
 import Category from '../../models/Category';
-import makeSafe from './make-safe';
 
 const router = Router();
 
 router.get('/', apiKey(), async (req, res) => {
-    makeSafe(res, async () => {
+    try {
         const categories = await Category.find().lean();
         res.json({categories});
-    });
+    }
+    catch (err) {
+        console.error(err);
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 router.get('/:_id', apiKey(), async (req, res) => {
-    makeSafe(res, async () => {
+    try {
         const _id = parseInt(req.params._id);
         if (!_id)
             res.status(400).json({error: BAD_ID});
@@ -30,11 +34,15 @@ router.get('/:_id', apiKey(), async (req, res) => {
             res.status(404).json({error: NOT_FOUND});
 
         res.json(category);
-    });
+    }
+    catch (err) {
+        console.error(err);
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 router.post('/', apiKey(), isAdmin(), json(), async (req, res) => {
-    makeSafe(res, async () => {
+    try {
         const {title} = req.body;
         const order = parseInt(req.body.order);
 
@@ -51,11 +59,15 @@ router.post('/', apiKey(), isAdmin(), json(), async (req, res) => {
             console.error(err);
             return res.status(400).json({error: INVALID});
         }
-    });
+    }
+    catch (err) {
+        console.error(err);
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 router.put('/:_id', apiKey(), isAdmin(), json(), async (req, res) => {
-    makeSafe(res, async () => {
+    try {
         const _id = parseInt(req.params._id);
         if (!_id)
             return res.status(404).json({error: BAD_ID});
@@ -80,16 +92,29 @@ router.put('/:_id', apiKey(), isAdmin(), json(), async (req, res) => {
             console.error(err);
             res.status(400).json({error: INVALID});
         }
-    })
+    }
+    catch (err) {
+        console.error(err);
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 router.delete('/:_id', apiKey(), isAdmin(), async (req, res) => {
-    const {_id} = req.query;
-    const result = await Category.findOneAndDelete({_id});
-    if (!result)
-        return res.status(404).json({error: NOT_FOUND});
+    try {
+        const _id = parseInt(req.params._id);
+        if (!_id)
+            return res.status(404).json({error: BAD_ID});
 
-    res.status(204).end();
+        const result = await Category.findOneAndDelete({_id});
+        if (!result)
+            return res.status(404).json({error: NOT_FOUND});
+
+        res.status(204).end();
+    }
+    catch (err) {
+        console.error(err);
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 export default router;
