+++ src/controllers/api/forums.ts
@@ -4,28 +4,31 @@ const NOT_FOUND = 'Forum not found.';
 const ALL_REQUIRED = 'Title, Order, and Category fields are required.';
 const SOME_REQUIRED = 'Title, Order, or Category field is required.';
 const INVALID = 'One or more fields are invalid.';
+const SERVER_ERROR = 'Server error.';
 
 import {Router, json} from 'express';
 import apiKey from '../../middleware/api-key';
-import makeSafe from './make-safe';
 import Forum from '../../models/Forum';
 import isAdmin from '../../middleware/is-admin';
 
 const router = Router();
 
 router.get('/', apiKey(), async (req, res) => {
-    makeSafe(res, async () => {
+    try {
         const {c: category} = req.query;
         if (!category)
             return res.status(400).json({error: CATEGORY_REQUIRED});
         
         const forums = await Forum.find({category}).lean();
         res.json({forums});
-    });
+    }
+    catch (err) {
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 router.get('/_id', apiKey(), isAdmin(), json(), async (req, res) => {
-    makeSafe(res, async () => {
+    try {
         const _id = parseInt(req.params._id);
         if (!_id)
             return res.status(400).json({error: BAD_ID});
@@ -35,11 +38,14 @@ router.get('/_id', apiKey(), isAdmin(), json(), async (req, res) => {
             return res.status(404).json({error: NOT_FOUND});
         
         res.json({forum});
-    });
+    }
+    catch (err) {
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 router.post('/', apiKey(), isAdmin(), json(), async (req, res) => {
-    makeSafe(res, async () => {
+    try {
         const {title} = req.body;
         const order = parseInt(req.body.order);
         const category = parseInt(req.body.category);
@@ -57,11 +63,14 @@ router.post('/', apiKey(), isAdmin(), json(), async (req, res) => {
             console.error(err);
             return res.status(400).json({error: INVALID});
         }
-    });
+    }
+    catch (err) {
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 router.put('/:_id', apiKey(), isAdmin(), json(), async (req, res) => {
-    makeSafe(res, async () => {
+    try {
         const _id = parseInt(req.params._id);
         if (!_id)
             return res.status(404).json({error: BAD_ID});
@@ -89,16 +98,27 @@ router.put('/:_id', apiKey(), isAdmin(), json(), async (req, res) => {
             console.error(err);
             res.status(400).json({error: INVALID});
         }
-    })
+    }
+    catch (err) {
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 router.delete('/:_id', apiKey(), isAdmin(), async (req, res) => {
-    const {_id} = req.query;
-    const result = await Forum.findOneAndDelete({_id});
-    if (!result)
-        return res.status(404).json({error: NOT_FOUND});
+    try {
+        const _id = parseInt(req.params._id);
+        if (!_id)
+            return res.status(404).json({error: BAD_ID});
 
-    res.status(204).end();
+        const result = await Forum.findOneAndDelete({_id});
+        if (!result)
+            return res.status(404).json({error: NOT_FOUND});
+
+        res.status(204).end();
+    }
+    catch (err) {
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 export default router;
