+++ src/controllers/api/subforums.ts
@@ -1,31 +1,35 @@
-const forum_REQUIRED = 'Forum ID is required.';
+const FORUM_REQUIRED = 'Forum ID is required.';
 const BAD_ID = 'Invalid ID.';
 const NOT_FOUND = 'subforum not found.';
 const ALL_REQUIRED = 'Title, Order, and Forum fields are required.';
 const SOME_REQUIRED = 'Title, Order, or Forum field is required.';
 const INVALID = 'One or more fields are invalid.';
+const SERVER_ERROR = 'Server error.';
 
 import {Router, json} from 'express';
 import apiKey from '../../middleware/api-key';
-import makeSafe from './make-safe';
 import isAdmin from '../../middleware/is-admin';
 import Subforum from '../../models/Subforum';
 
 const router = Router();
 
 router.get('/', apiKey(), async (req, res) => {
-    makeSafe(res, async () => {
+    try {
         const {f: forum} = req.query;
         if (!forum)
-            return res.status(400).json({error: forum_REQUIRED});
+            return res.status(400).json({error: FORUM_REQUIRED});
         
         const subforums = await Subforum.find({forum}).lean();
         res.json({subforums});
-    });
+    }
+    catch (err) {
+        console.error(err);
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 router.get('/_id', apiKey(), isAdmin(), json(), async (req, res) => {
-    makeSafe(res, async () => {
+    try {
         const _id = parseInt(req.params._id);
         if (!_id)
             return res.status(400).json({error: BAD_ID});
@@ -35,11 +39,15 @@ router.get('/_id', apiKey(), isAdmin(), json(), async (req, res) => {
             return res.status(404).json({error: NOT_FOUND});
         
         res.json({subforum});
-    });
+    }
+    catch (err) {
+        console.error(err);
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 router.post('/', apiKey(), isAdmin(), json(), async (req, res) => {
-    makeSafe(res, async () => {
+    try {
         const {title} = req.body;
         const order = parseInt(req.body.order);
         const forum = parseInt(req.body.forum);
@@ -57,11 +65,15 @@ router.post('/', apiKey(), isAdmin(), json(), async (req, res) => {
             console.error(err);
             return res.status(400).json({error: INVALID});
         }
-    });
+    }
+    catch (err) {
+        console.error(err);
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 router.put('/:_id', apiKey(), isAdmin(), json(), async (req, res) => {
-    makeSafe(res, async () => {
+    try {
         const _id = parseInt(req.params._id);
         if (!_id)
             return res.status(404).json({error: BAD_ID});
@@ -89,16 +101,29 @@ router.put('/:_id', apiKey(), isAdmin(), json(), async (req, res) => {
             console.error(err);
             res.status(400).json({error: INVALID});
         }
-    })
+    }
+    catch (err) {
+        console.error(err);
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 router.delete('/:_id', apiKey(), isAdmin(), async (req, res) => {
-    const {_id} = req.query;
-    const result = await Subforum.findOneAndDelete({_id});
-    if (!result)
-        return res.status(404).json({error: NOT_FOUND});
+    try {
+        const _id = parseInt(req.params._id);
+        if (!_id)
+            return res.status(404).json({error: BAD_ID});
 
-    res.status(204).end();
+        const result = await Subforum.findOneAndDelete({_id});
+        if (!result)
+            return res.status(404).json({error: NOT_FOUND});
+
+        res.status(204).end();
+    }
+    catch (err) {
+        console.error(err);
+        res.status(500).json({error: SERVER_ERROR});
+    }
 });
 
 export default router;
